Restauration complete du prix initial

+++ shop/models.py
@@ -1,7 +1,7 @@
 from django.db import models
 
 class Categorie(models.Model):
-    nom = models.CharField(max_length=100)  # Corrigé ici
+    nom = models.CharField(max_length=100)
     slug = models.SlugField(unique=True)
 
     def __str__(self):
@@ -9,9 +9,9 @@ class Categorie(models.Model):
 
 class Produit(models.Model):
     categorie = models.ForeignKey(Categorie, on_delete=models.CASCADE, related_name='produits')
-    nom = models.CharField(max_length=200)  # Corrigé ici
+    nom = models.CharField(max_length=200)
     description = models.TextField(blank=True)
-    prix = models.IntegerField()
+    prix = models.DecimalField(max_digits=10, decimal_places=2)  # Remis comme au début
     stock = models.IntegerField(default=0)
     disponible = models.BooleanField(default=True)
     image = models.ImageField(upload_to='produits/', blank=True, null=True)
